Video & decoding changes:
* Offload mini preview from decoding the actual camera QR codes to new 
hidden canvas. This way decoding area's resolution are not affected by 
the size of the mini preview.

* Handle Video slot on each decoding and pass it along few events. Also 
handle Video dimensions for each of the two Videos. This all is to adapt 
properly to the multi camera scenario.

* Fix JOB not successfully reading codes (authentication photo uploads). 
Video decoding interrumped it since JOB is asynchronous. Added a mutex 
to solve it.

* Refactor, other...